\name{align_ssb_r}
\alias{align_ssb_r}
\title{Align stock and recruitment data}
\usage{
align_ssb_r(data, col.names = c("year", "r", "ssb"), aR)
}
\arguments{
  \item{data}{data.frame that contains in its first three
  rows year, recruitment and ssb.}

  \item{col.names}{vector that contains the names for year,
  recruitment and ssb.}

  \item{aR}{integer that contains the age of recruits}
}
\description{
XXX
}

