\name{bioIndex.Original}
\alias{bioIndex.Original}
\title{bioIndex.Original}
\usage{
bioIndex.Original(lengths, stations, KYN, lengdir, yr)
}
\arguments{
  \item{lengths}{Data frame containing length information}

  \item{stations}{Data frame containg station information}

  \item{KYN}{kyn}

  \item{lengdir}{Lengdir}

  \item{yr}{Year}
}
\description{
Original long stuff in SMB BIOVISIT_R shell script. Runs in
a loop for each length class
}

