\name{cat_srmsymc.dat}
\alias{cat_srmsymc.dat}
\title{Write srmsymc.dat to disk}
\usage{
  cat_srmsymc.dat(name_stock, filename_age, y1, y2, aR, aP,
    opt_sr_model, opt_sim, opt_age, opt_pen, r, ssb, year)
}
\arguments{
  \item{name_stock}{XXX}

  \item{filename_age}{XXX}

  \item{y1}{XXX}

  \item{y2}{XXX}

  \item{aR}{XXX}

  \item{aP}{XXX}

  \item{opt_sr_model}{XXX}

  \item{opt_sim}{XXX}

  \item{opt_age}{XXX}

  \item{opt_pen}{XXX}

  \item{r}{XXX}

  \item{ssb}{XXX}

  \item{year}{XXX}
}
\description{
  XXX
}
\author{
  Einar Hjorleifsson
}

