\name{Combine.strata}
\alias{Combine.strata}
\alias{index.aggregate.combine}
\title{Sums mean, standard error, cv by aggregated area}
\usage{
Combine.strata(result, index)
}
\arguments{
  \item{result}{Result}

  \item{index}{Index}
}
\description{
Function used both in \code{\link{Calc.index}} and
\code{\link{index.aggregate}}.
}

