\name{compile_catch}
\alias{compile_catch}
\title{compile_catch}
\usage{
compile_catch(species = "torskur", tegund = 1, ar = 2012)
}
\arguments{
  \item{species}{character vector containing species name}

  \item{tegund}{numeric value containing specise number}

  \item{ar}{year}
}
\description{
Compiles logbook catches
}

