\name{contour_plot3}
\alias{contour_plot3}
\title{Contour plot}
\usage{
contour_plot3(logs, LEVELS = c(1, 5, 10, 15, 20, 40), colour = c(0,
  brewer.pal(5, "YlOrRd"), "black"), plotit = FALSE, geopar)

contour_plot3(logs, LEVELS = c(1, 5, 10, 15, 20, 40), colour = c(0,
  brewer.pal(5, "YlOrRd"), "black"), plotit = FALSE, geopar)
}
\arguments{
  \item{logs}{XXX}

  \item{LEVELS}{XXX}

  \item{colour}{XXX}

  \item{plotit}{XXX}

  \item{logs}{XXX}

  \item{LEVELS}{XXX}

  \item{colour}{XXX}

  \item{plotit}{XXX}

  \item{geopar}{XXX}
}
\description{
Makes contour plots

Makes contour plots
}

