\name{contour_plot4}
\alias{contour_plot4}
\title{Contour plot}
\usage{
contour_plot4(logs, Levels = c(1, 5, 10, 15, 20, 40), Colour = c(0,
  brewer.pal(5, "YlOrRd"), "black"), plotit = FALSE)
}
\arguments{
  \item{logs}{XXX}

  \item{Levels}{XXX}

  \item{Colour}{XXX}

  \item{plotit}{XXX}
}
\description{
Makes contour plots
}

