\name{date_2_qyear_start}
\alias{date_2_qyear_start}
\title{Converts date to first date (1. september) of each quota year}
\usage{
  date_2_qyear_start(x)
}
\arguments{
  \item{x}{A vector containing dates}
}
\description{
  Note that the first quota year was from 1. januar 1991 to
  31. august 1991. Taken care of in the function.
}

