\docType{package}
\name{fishvise-package}
\alias{fishvise}
\alias{fishvise-package}
\alias{fishviseLibrary}
\title{Title Fishvice package}
\description{
  More about what it does (maybe more than one line)
}
\details{
  Version 1.0: 14 December 2010
}
\author{
  Various folks at MRI
}
\seealso{
  link to other documents
}

