% Generated by roxygen2 (4.0.1): do not edit by hand
\name{FLStock2rbya}
\alias{FLStock2rbya}
\title{Convert FLStock (or FLSAM) object long data.frame}
\usage{
FLStock2rbya(x)
}
\arguments{
\item{x}{FLStock or FLSAM object}
}
\value{
A data.frame containing the fields: year, age,
n (stock in numbers),
f (fishing mortality),
oC (observed catch),
cW (catch weights),
ssbW (spawning stock weights),
oD (observed dicards),
dW (discard weights),
oL (observed landings),
lW (landings weights),
mat (maturity),
pF (partial fishing mortality before spawning)
pM (partial natural mortality before spawning),
m (natural mortality)
and f (fishing mortality).
}
\description{
Converts FLStock or FLSAM object to rbya (Results By Year and Age) format.
}
\note{
TO DO: FLSAM only contains the stock in numbers and fishing mortality by
year and age(?????). Need to "adjust" the function to accomodate that.
}

