% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_sam}
\alias{get_sam}
\title{get_sam}
\usage{
get_sam(assessment = "wbcod_2014", user = "user3")
}
\arguments{
\item{assessment}{The directory name of the assessment as specified on
www.stockassessment.org}

\item{user}{User name, guest users can use "user3" (default)}
}
\description{
Gets the input files and sam output files from the www.stockassessment.org
}

