% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gJoin}
\alias{gJoin}
\title{gJoin}
\usage{
gJoin(x, y, column, name.x, name.y, outer.join = T, set = NA)
}
\arguments{
\item{x}{X}

\item{y}{Y}

\item{column}{Column}

\item{name.x}{nameX}

\item{name.y}{nameY}

\item{outer.join}{outerJoin}

\item{set}{Set}
}
\description{
The old join from geo, renamed to avoid confusion with join from plyr
}

