\name{hcr_observation_error}
\alias{hcr_observation_error}
\title{Observation error model}
\usage{
  hcr_observation_error(d, y, h, hrate, Fmult, delay,
    ass_bias, ass_error_type)
}
\arguments{
  \item{d}{XXX}

  \item{y}{XXX}

  \item{h}{XXX}

  \item{hrate}{The harvest rate (fishing mortality)}

  \item{Fmult}{XXX}

  \item{delay}{XXX}

  \item{ass_bias}{A constant}

  \item{ass_error_type}{XXX}
}
\description{
  XXX
}

