% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hcr_operating_model2}
\alias{hcr_operating_model2}
\title{Operating model2}
\usage{
hcr_operating_model2(y, h, ctr, Fmult, nR = 1)
}
\arguments{
\item{y}{XXX}

\item{h}{XXX}

\item{ctr}{ctr_rec}

\item{Fmult}{XXX}

\item{nR}{XXX}
}
\description{
This function is the same as hcr_operating_model except
the recruitment model
}

