\name{hcr_set_starting_conditions}
\alias{hcr_set_starting_conditions}
\title{HCR: Set starting condition for stock}
\usage{
  hcr_set_starting_conditions(dat_y1, d, ctr,
    wgt_error_type, refBweight)
}
\arguments{
  \item{dat_y1}{XXX}

  \item{d}{XXX}

  \item{ctr}{XXX}

  \item{wgt_error_type}{XXX}

  \item{refBweight}{XXX}
}
\description{
  XXX
}

