\docType{data}
\name{LENGDIR}
\alias{LENGDIR}
\title{Some title}
\format{Some format}
\source{
  \code{ attach('/net/hafkaldi/u2/reikn/Splus5/SMB/.RData')
  save('LENGDIR',file='data/LENGDIR.rda')
  detach('file:/net/hafkaldi/u2/reikn/Splus5/SMB/.RData')}
}
\description{
  from SMB RData
}
\author{
  Some author
}
\keyword{datasets}

