\docType{data}
\name{LWCOEFF}
\alias{LWCOEFF}
\title{Length weigth coefficients}
\format{Each element of the list contains parameter a and b for different
species. The name of the element of the list corresponds to numerical species
code.}
\source{
  The object is created and maintained by Höskuldur
  Björnsson. The original source is found in
  "/net/hafkaldi/u2/reikn/Splus5/SMB/.RData". It is copied
  into this library via: \code{
  attach('/net/hafkaldi/u2/reikn/Splus5/SMB/.RData')
  save('LWCOEFF',file='data/LWCOEFF.rda')
  detach('file:/net/hafkaldi/u2/reikn/Splus5/SMB/.RData')}
}
\description{
  A \emph{list} that contains length-weight coefficients
  for number of species.
}
\author{
  Hoskuldur Bjornsson
}
\keyword{datasets}

