\name{mb_sql}
\alias{mb_sql}
\title{SQL function for reading SQLite database}
\usage{
  mb_sql(query,
    dbname = "/net/hafkaldi/u1/mb/prj/2013_11topovise/dbase/mb.sqlite")
}
\arguments{
  \item{query}{The sql query}

  \item{dbname}{The SQLite database to connect to}
}
\description{
  A little helper function for reading the mb-database
}

