% Generated by roxygen2 (4.0.1): do not edit by hand
\name{process_error_sam}
\alias{process_error_sam}
\title{process_error_sam}
\usage{
process_error_sam(rbya)
}
\arguments{
\item{rbya}{data.frame containing stock in numbers (n), fishing mortality (f) and
natural mortality (m) by year (year) and age (age).}
}
\description{
Calculates the sam process error
}

