\name{qCounted}
\alias{qCounted}
\title{qCounted}
\usage{
qCounted(Lengths, Numbers, col.names = c("fj.maelt", "fj.talid"),
  id.name = c("synis.id"))
}
\arguments{
  \item{Lengths}{data.frame containing lengths}

  \item{Numbers}{data.frame containing numbers}

  \item{col.names}{A character vector containing field name
  of n lengths, and n counted}

  \item{id.name}{the name that joins Numbers to Lengths,
  default synis.id}
}
\description{
What it does
}

