\name{r2d}
\alias{r2d}
\title{Converts statistical squares to degrees}
\usage{
  r2d(square)
}
\arguments{
  \item{square}{input vector, numerical}
}
\value{
  a list with lat and lon
}
\description{
  \code{r2d} converts statistical rectangles to degrees
}
\author{
  Hoski, documented by Einar
}
\seealso{
  \code{\link{sr2d}} for converting 'small' statistical
  squares
}

