% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Rattach}
\alias{Rattach}
\title{Rattach}
\usage{
Rattach(what, pos, ...)
}
\arguments{
\item{what}{The directory containing .RData}

\item{pos}{integer specifying position in search() where to attach.
If not specified, will attached to search path 2.}

\item{...}{other arguments to pass to \code{attach}}
}
\description{
Modified \code{attach}. If in search path is, unless otherwise
specied put in search path 2.
}

