% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.lowestoft}
\alias{read.lowestoft}
\title{Read in Lowestoft-format VPA data}
\usage{
read.lowestoft(filename, Format = "List")
}
\arguments{
\item{filename}{The name of the file to read in}

\item{Format}{The format of the output, available is "List","Wide","Long"}
}
\description{
Read in Lowestoft-format VPA data
}

