\name{read.prelude}
\alias{read.prelude}
\title{Reads prelude-files}
\usage{
read.prelude(skr, rownames = F, dots.in.text = T)
}
\arguments{
  \item{skr}{is the name of the file}

  \item{rownames,}{default is FALSE}

  \item{dots.in.text,}{default is TRUE}
}
\description{
\code{read.prelude} reads prelude files. Originally located
on package geo, there called s2pre.
}
\author{
Hoski, documented by Einar
}

