\name{read.sen}
\alias{read.sen}
\title{Read .sen file}
\usage{
read.sen(file, pM, pF)
}
\arguments{
  \item{file}{Name of the .sen file}

  \item{pM}{vector containing proportional natural
  mortality before spawning}

  \item{pF}{vector containing proportional fishing
  mortality before spawning}
}
\description{
Reads a file of the .sen format (Aberdeen format) and
creates a \code{data.frame}. Additional informations are
stored as attributes.
}
\author{
Einar Horleifsson <einar.hjorleifsson
}

