% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_conf_sam}
\alias{read_conf_sam}
\title{read_conf_sam}
\usage{
read_conf_sam(clonefile)
}
\arguments{
\item{clonefile}{XXX}
}
\description{
Reads sam configuration file from the stockassessment.org
directory structure.

Used by read.fit function (see below). This function originally in
scr/common.R named read.conf
}

