% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_fit_sam}
\alias{read_fit_sam}
\title{read_fit_sam}
\usage{
read_fit_sam(file, reduced = FALSE)
}
\arguments{
\item{file}{Path to the directory containing sam run. Normally ends with
../run/sam}

\item{reduced}{Boolean, if FALSE (default) ....}
}
\value{
Returns a list
}
\description{
Reads the stock assessment summary statistics from the
stockassessment.org directory structure

This function originally in scr/common.R as read.fit
}
\note{
TO DO: Ideally should convert the data to FLStock object and from there
convert the stuff to e.g. rby or rbya.

TO DO: Provide detail description of the list returned
}

