% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_lowestoft}
\alias{read_lowestoft}
\title{Read in Lowestoft-format VPA data}
\usage{
read_lowestoft(filename, val.name, Format = "List")
}
\arguments{
\item{filename}{The name of the file to read in}

\item{val.name}{XXX}

\item{Format}{The format of the output, available is "List","Wide","Long"}
}
\description{
Read in Lowestoft-format VPA data
}

