% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_lowestoft2}
\alias{read_lowestoft2}
\title{read_lowestoft2}
\usage{
read_lowestoft2(file, sep = "", quiet = TRUE)
}
\arguments{
\item{file}{name of file, normally the index file name}

\item{sep}{the separator, default is ""}

\item{quiet}{boolean, default is TRUE}
}
\description{
Modified \href{https://github.com/flr/FLCore/blob/master/R/io.VPAsuite.R}{FLCore::readVPA}
function in the FLCore-package, except here it only returns a list or a data.frame
}

