% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_lowestoft_file}
\alias{read_lowestoft_file}
\title{read_lowestoft_file}
\usage{
read_lowestoft_file(file, sep = "", quiet = TRUE)
}
\arguments{
\item{file}{name of file, normally the index file name}

\item{sep}{the separator, default is ""}

\item{quiet}{boolean, default is TRUE}
}
\description{
bla, bla, ... Modified
\href{https://github.com/flr/FLCore/blob/master/R/io.VPAsuite.R}{FLCore::readVPAFile} function,
here it only returns an matrix rather than a FLQuant-object
}

