% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_lowestoft_survey}
\alias{read_lowestoft_survey}
\title{Read in Lowestoft-format survey data}
\usage{
read_lowestoft_survey(filename)
}
\arguments{
\item{filename}{The name of the file to read in}
}
\description{
Read in Lowestoft-format survey data
}

