% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_rby_sam}
\alias{read_rby_sam}
\title{read_rby_sam}
\usage{
read_rby_sam(x, range = FALSE, Scale = 1, format = "wide",
  stdev = FALSE)
}
\arguments{
\item{x}{Object from function \code{read.fit}.}

\item{range}{Boolean Not used}

\item{Scale}{A value}

\item{format}{Character string. If "wide" (default) returns a wide table.
If "long", returns a data.frame containing year, variable and est, hig and low
for each}

\item{stdev}{Boolean, if FALSE (default) not included in the output}
}
\description{
Reads the stock summary numbers by year from \code{sam}.
}

