% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_rbya_sam}
\alias{read_rbya_sam}
\title{read_rbya_sam}
\usage{
read_rbya_sam(fit, Scale = 1)
}
\arguments{
\item{fit}{Object from read_fit_sam}

\item{Scale}{A value}
}
\description{
read fishing mortality and stock in numbers from sam
}

