% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_sam}
\alias{read_sam}
\title{read_sam}
\usage{
read_sam(run.dir, ctrl, admb.stem = "sam", ibya)
}
\arguments{
\item{run.dir}{Path to which the files are read from.}

\item{ctrl}{If missing (default) the relevant information are obtained
from the files in the run.dir. If specified this must be a FLSAM.control
object containing the parameters employed in the assessment.}

\item{admb.stem}{The name of the admb executable (ie without the .dat, .exe etc)}

\item{ibya}{A long data.frame containing input values. If missing (default)
the returned rbya data.frame does not contain the full input data.}
}
\description{
Reads results from sam. Equivalent to FLSAM::SAM2FLR except
returns a list rather than an FLR-object.
}

