% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_sen}
\alias{read_sen}
\title{Read .sen file}
\usage{
read_sen(file, pM, pF)
}
\arguments{
\item{file}{Name of the .sen file}

\item{pM}{vector containing proportional natural mortality before spawning}

\item{pF}{vector containing proportional fishing mortality before spawning}
}
\description{
Reads a file of the .sen format (Aberdeen format) and creates a
\code{data.frame}. Additional informations are stored as
attributes.
}
\author{
Einar Horleifsson <einar.hjorleifsson@gmail.com>
}

