% Generated by roxygen2 (4.0.1): do not edit by hand
\name{replace.text}
\alias{replace.text}
\title{Replace characters}
\usage{
replace.text(txt, char = "_", replchar = ".")
}
\arguments{
\item{txt}{input vector}

\item{char}{character to be replaced.}

\item{replchar}{replacement character.}
}
\description{
\code{replace.text} replaces string with another string. Function was called 'skipta.texta'. In theory could use str_replace instead.
}

