% Generated by roxygen2 (4.0.1): do not edit by hand
\name{residplot}
\alias{residplot}
\title{residplot}
\usage{
residplot(residuals, xpos, ypos, maxsize = 0.2, poscol = 2, linecol = 1,
  lwd = 1, n = 50, maxn, negcol, txt = F, cex = 0.5, xlab = "",
  ylab = "", axes = T, arg = T, argcol = 20, arglty = 2, cn = c("x",
  "y", "z"), append = F)
}
\arguments{
\item{residuals}{XXX}

\item{xpos}{XXX}

\item{ypos}{XXX}

\item{maxsize}{XXX}

\item{poscol}{XXX}

\item{linecol}{XXX}

\item{lwd}{XXX}

\item{n}{XXX}

\item{maxn}{XXX}

\item{negcol}{XXX}

\item{txt}{XXX}

\item{cex}{XXX}

\item{xlab}{XXX}

\item{ylab}{XXX}

\item{axes}{XXX}

\item{arg}{XXX}

\item{argcol}{XXX}

\item{arglty}{XXX}

\item{cn}{XXX}

\item{append}{XXX}
}
\description{
A bubble plot of residuals
}
\author{
Hoskuldur Bjornsson
}

