\name{rvk2flr}
\alias{rvk2flr}
\title{Convert reykjavik to flr}
\usage{
rvk2flr(rbx, sName = "nn", sDesc = "none", pf = 0, pm = 0, cF1 = 5,
  cF2 = 10)
}
\arguments{
  \item{rbx}{rbx object}

  \item{sName}{character. Name of stock}

  \item{sDesc}{character. Some description}

  \item{pf}{XX}

  \item{pm}{XX}

  \item{cF1}{XX}

  \item{cF2}{XX}
}
\description{
xx
}

