% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rvk2FLStock}
\alias{rvk2FLStock}
\title{Convert rvk-format to FLStock object}
\usage{
rvk2FLStock(rbx, sName = "nn", sDesc = "none", pf = 0, pm = 0,
  aF1 = 5, aF2 = 10)
}
\arguments{
\item{rbx}{A list with two data.frames named rby and rbya.}

\item{sName}{Character vector containing name of stock}

\item{sDesc}{Character vectro containing some description}

\item{pf}{Partial fishing mortality}

\item{pm}{Partial natural mortality}

\item{aF1}{Lower reference age}

\item{aF2}{Upper reference age}
}
\description{
Converts the rby and rbya to FLStock object.
}
\note{
TO DO: check the unit stuff
}

