\name{sr2d}
\alias{sr2d}
\title{Converts small statistical squares to degrees}
\usage{
sr2d(ssquare)
}
\arguments{
  \item{ssquare}{input vector, numerical}
}
\value{
a list with lat and lon
}
\description{
\code{sr2d} converts small statistical rectangles to
degrees
}
\author{
Hoski, documented by Einar
}
\seealso{
\code{\link{r2d}} for converting statistical squares to lat
and lon
}

