\docType{data}
\name{Std.aggregation}
\alias{Std.aggregation}
\title{Some title}
\format{Some format}
\source{
  The object is created and maintained by Höskuldur
  Björnsson. The original source is found in
  "/net/hafkaldi/u2/reikn/Splus5/SMB/.RData". It is copied
  into this library via: \code{
  attach("/net/hafkaldi/u2/reikn/Splus5/SMB/GEOMETRY.NEW/.RData")
  names(Std.aggregation) <-
  iconv(names(Std.aggregation),'ISO8859-1','UTF-8')
  save(Std.aggregation,file='data/Std.aggregation.rda')
  detach('file:/net/hafkaldi/u2/reikn/Splus5/SMB/.RData')}
}
\description{
  from SMB RData
}
\author{
  Some author
}
\keyword{datasets}

