% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trim_shape_df}
\alias{trim_shape_df}
\title{trim_shape_df}
\usage{
trim_shape_df(x, xlim = c(-30, 10), ylim = c(64, 67))
}
\arguments{
\item{x}{A shape data frame}

\item{xlim}{Longitudinal boundaries}

\item{ylim}{Latitudinal boundaries}
}
\description{
Trims shape data.frames and returns a shape data.frame
}

