% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wget_sam_run}
\alias{wget_sam_run}
\title{wget_sam_run}
\usage{
wget_sam_run(dir)
}
\arguments{
\item{dir}{Name of the directory on stockassessment.org}
}
\description{
Fetches the results from the run directory on stockassessment.org
and stored it locally.
}
\note{
#' Currently the directory structure is also copied, see datadisk/stock...

TO DO: Get rid of the long directory path in the stuff returned to the
local computer.
}

