% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ypr_single_target}
\alias{ypr_single_target}
\title{ypr}
\usage{
ypr_single_target(fmort, cW, selC, selL, M, ages = c(2:9),
  plusgroup = plusgroup, mat, pM = 0, pF = 0, dW, ssbW)
}
\arguments{
\item{fmort}{A vector containing the fishing mortality over which to run the ypr}

\item{cW}{Catch weights}

\item{selC}{Selection pattern of the catch}

\item{selL}{Selection pattern of the landings}

\item{M}{Natural mortality, can be a single value}

\item{ages}{A vector containing the age classes}

\item{plusgroup}{Boolean If FALSE (default) no plusgroup bookkeeping}

\item{mat}{A vector of proportion mature}

\item{pM}{Partial natural mortality prior to spawning}

\item{pF}{Partial fishing mortality prior to spawning}

\item{dW}{Discard weights}

\item{ssbW}{Spawning stock weights}
}
\description{
A yield per recruit model
}

