% Generated by roxygen2 (4.0.1): do not edit by hand
\name{apply.shrink}
\alias{apply.shrink}
\title{apply.shrink}
\usage{
apply.shrink(X, INDICES, FUN = NULL, names, ...)
}
\arguments{
\item{X}{x}

\item{INDICES}{indices}

\item{FUN}{fun}

\item{names}{Names}

\item{...}{additionalThings}
}
\description{
local copy of an old faithful from package geo
}

